Refactor MaxQuerySizeExceeded.tsx
